Improve error handling (#14)
* Implement Outcome functor

This commit adds the `Outcome` functor that would be used in the project to handle errors in code. This class is used for educational purposes only, and would be replaced later on by Either from ArrowKT library

* Refactor routing functions

This commit refactoring the routing functions to be more accurate and easy to read. Moreover it updates the HTML pages of a single list and lists of ToDo's list

* Convert extract methods to user Outcome

This commit changes the extract methods on the routing to use Outcome class instead of nullability for a better error handling

* Convert hub to use Outcome

This commit converts the hub to user the Outcome class, and updates the calling path accordingly

* Adjust tests

This commit adjust the tests to pass after the refactoring

---------

Co-authored-by: Yonatan Karp-Rudin <yonatan.karp@sumup.com>